Nav-bar fixed to upper right
Re-did nav-bar to keep it visible while scrolling.
Changed the logo and tile flex container to justify center to keep them together and centered.
Added another product description to improve the layout.
Added another justify attribute to the CSS for the product flex container to allow them to space out.