Addresses Issue #8
Addresses Issue #8.
Made a solid bar across the screen for the navigation menu. Changed the background color of the menu items and improved the hover action.
The highlight bar in the product box does not consistently stay aligned. Replaced with a linear gradient stripe.
Cleaned up CSS naming a little